+++ js/src/jit/mips-shared/CodeGenerator-mips-shared.cpp
@@ -297,7 +297,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
     }
 
     if (constant > 0) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uint32_t(constant));
 
       if (!mul->canOverflow()) {
         
@@ -397,7 +397,7 @@ void CodeGeneratorMIPSShared::emitMulI64(Register lhs, int64_t rhs,
 
   if (rhs > 0) {
     if (std::has_single_bit(static_cast<uint64_t>(rhs + 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs + 1);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs + 1));
 
       UseScratchRegisterScope temps(masm);
       Register savedLhs = lhs;
@@ -411,7 +411,7 @@ void CodeGeneratorMIPSShared::emitMulI64(Register lhs, int64_t rhs,
     }
 
     if (std::has_single_bit(static_cast<uint64_t>(rhs - 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs - 1u);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs - 1u));
 
       UseScratchRegisterScope temps(masm);
       Register savedLhs = lhs;
@@ -425,7 +425,7 @@ void CodeGeneratorMIPSShared::emitMulI64(Register lhs, int64_t rhs,
     }
 
     
-    int32_t shift = mozilla::FloorLog2(rhs);
+    int32_t shift = mozilla::FloorLog2(uint64_t(rhs));
     if (int64_t(1) << shift == rhs) {
       masm.lshiftPtr(Imm32(shift), lhs, dest);
       return;
