+++ js/src/jit/riscv64/CodeGenerator-riscv64.cpp
@@ -841,7 +841,7 @@ void CodeGenerator::visitMulI(LMulI* ins) {
     }
 
     if (constant > 0) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uint32_t(constant));
 
       if (!mul->canOverflow()) {
         
@@ -943,7 +943,7 @@ void CodeGeneratorRiscv64::emitMulI64(Register lhs, int64_t rhs,
 
   if (rhs > 0) {
     if (std::has_single_bit(static_cast<uint64_t>(rhs + 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs + 1);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs + 1));
 
       UseScratchRegisterScope temps(&masm);
       Register savedLhs = lhs;
@@ -957,7 +957,7 @@ void CodeGeneratorRiscv64::emitMulI64(Register lhs, int64_t rhs,
     }
 
     if (std::has_single_bit(static_cast<uint64_t>(rhs - 1))) {
-      int32_t shift = mozilla::FloorLog2(rhs - 1);
+      int32_t shift = mozilla::FloorLog2(uint64_t(rhs - 1));
 
       UseScratchRegisterScope temps(&masm);
       Register savedLhs = lhs;
@@ -971,7 +971,7 @@ void CodeGeneratorRiscv64::emitMulI64(Register lhs, int64_t rhs,
     }
 
     
-    uint8_t shift = mozilla::FloorLog2(rhs);
+    uint8_t shift = mozilla::FloorLog2(uint64_t(rhs));
     if (int64_t(1) << shift == rhs) {
       masm.slli(dest, lhs, shift);
       return;
