+++ js/src/jit/x64/CodeGenerator-x64.cpp
@@ -184,7 +184,7 @@ void CodeGenerator::visitMulI64(LMulI64* lir) {
         break;
       default: {
         
-        int32_t shift = mozilla::FloorLog2(constant);
+        int32_t shift = mozilla::FloorLog2(uint64_t(constant));
         if (constant > 0 && (1 << shift) == constant) {
           if (lhs != out) {
             masm.movq(lhs, out);
@@ -1504,7 +1504,7 @@ void CodeGenerator::visitMulIntPtr(LMulIntPtr* ins) {
 
     
     if (constant > 0 && std::has_single_bit(uintptr_t(constant))) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uintptr_t(constant));
       masm.lshiftPtr(Imm32(shift), lhs);
       return;
     }
